progress on API, filtering, projection, plotting #13
