+++ orc/__init__.py
@@ -1,4 +1,4 @@
-from orc import io, cv, log, piv_process, frames
+from orc import io, cv, log, piv_process, frames, helpers, piv
 from orc.api import *
 
 # from orc import piv
